moved text measurements to new file
